+++ src/shell.rs
@@ -1,3 +1,5 @@
 //! Embedded hook text.
 
 pub const ZSH_HOOK: &str = include_str!("../shell/init.zsh");
+pub const BASH_HOOK: &str = include_str!("../shell/init.bash");
+pub const FISH_HOOK: &str = include_str!("../shell/init.fish");